LayoutTests/svg/dynamic-updates/SVGFEColorMatrixElement-svgdom-type-prop-expected.txt

CONSOLE MESSAGE: line 172: ReferenceError: Can't find variable: SVGFEColorMatrixElement
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Tests dynamic updates of the 'type' property of the SVGFEColorMatrixElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS matrixElement.getAttribute('type') is "saturate"
PASS saturateElement.getAttribute('type') is "hueRotate"
PASS hueRotateElement.getAttribute('type') is "luminanceToAlpha"
PASS luminanceToAlphaElement.getAttribute('type') is "matrix"
